pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "px4-sitl-doctor"
version = "0.1.0"
description = "Pre-launch environment validator for the PX4 + ROS 2 + Gazebo SITL stack"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["px4", "ros2", "gazebo", "sitl", "drone", "uav", "robotics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "packaging>=23.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
]

[project.scripts]
px4-doctor = "px4_doctor.cli:main"

[project.urls]
Homepage = "https://github.com/carloangulo/px4-sitl-doctor"
Repository = "https://github.com/carloangulo/px4-sitl-doctor"
Issues = "https://github.com/carloangulo/px4-sitl-doctor/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["px4_doctor*"]

[tool.setuptools.package-data]
px4_doctor = ["data/*.yaml"]
